project_name/frontend/requirements.txt
streamlit
numpy
matplotlib
Pillow
torch
torchvision
torchaudio 
Flask
flask-cors
requests